+++ src/main/java/com/example/campus/controller/TopController.java
@@ -1,0 +1,22 @@
+package com.example.campus.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+
+@Controller
+public class TopController {
+
+	//Get
+	@GetMapping("/top")
+	public String top() {
+		return "top";
+	}
+
+	//Post
+	@PostMapping("/top")
+	public String postTop() {
+		return "top";
+	}
+
+}
