+++ src/main/java/com/example/campus/repository/QuestionRepository.java
@@ -1,8 +1,10 @@
-package com.example.campus.model;
+package com.example.campus.repository;
 
 import java.util.List;
 
 import org.apache.ibatis.annotations.Mapper;
+
+import com.example.campus.entity.Question;
 
 
 @Mapper
