+++ src/main/java/com/example/campus/service/QuestionService.java
@@ -1,9 +1,12 @@
-package com.example.campus.model;
+package com.example.campus.service;
 
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import com.example.campus.entity.Question;
+import com.example.campus.repository.QuestionRepository;
 
 @Service //SpringのDIコンテナの管理対象であることを示す
 public class QuestionService {
