--- src/main/resources/com/example/campus/auth/UserRepository.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC
-	"-//mybatis.org//DTD Mapper 3.0//EN"
-	"http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<!-- namespace：mapperインターフェースのパッケージ付きクラス名を指定 -->
-<mapper namespace="com.example.campus.auth.UserRepository">
-
-	<!--今回利用するステートメントの定義 -->
-	<select id = "identifyUser" parameterType="java.lang.String" resultMap="userDetails">
-	select
-		id, password
-		from users
-		where id = #{id}
-		and deleteflag = 0
-	</select>
-
-	<!--
-	javaオブジェクトとsqlの実行結果をマッピングするためのもの
-	type属性にはメソッドの戻り値を指定する＝Userクラス（UserDetails インターフェースの実装クラス）
-	id属性は任意。selectで使う。
-	-->
-	<resultMap type="com.example.campus.auth.User" id ="userDetails">
-		<result property="id" column="id" />
-		<result property="password" column="password" />
-	</resultMap>
-</mapper>
